[process-agent] Move `setHostMountEnv` to `pkg/config`
